move PANIC_REASONS to transactions package
