--- packages/program/src/configs.ts
@@ -1,51 +1,0 @@
-// From https://github.com/FuelLabs/fuel-asm/blob/eb78378c3b7c22a53b834381c387d89b3c0ef122/src/panic_reason.rs#L13
-export const PANIC_REASONS = [
-  'UnknownPanicReason',
-  'Revert',
-  'OutOfGas',
-  'TransactionValidity',
-  'MemoryOverflow',
-  'ArithmeticOverflow',
-  'ContractNotFound',
-  'MemoryOwnership',
-  'NotEnoughBalance',
-  'ExpectedInternalContext',
-  'AssetIdNotFound',
-  'InputNotFound',
-  'OutputNotFound',
-  'WitnessNotFound',
-  'TransactionMaturity',
-  'InvalidMetadataIdentifier',
-  'MalformedCallStructure',
-  'ReservedRegisterNotWritable',
-  'InvalidFlags',
-  'InvalidImmediateValue',
-  'ExpectedCoinInput',
-  'EcalError',
-  'MemoryWriteOverlap',
-  'ContractNotInInputs',
-  'InternalBalanceOverflow',
-  'ContractMaxSize',
-  'ExpectedUnallocatedStack',
-  'MaxStaticContractsReached',
-  'TransferAmountCannotBeZero',
-  'ExpectedOutputVariable',
-  'ExpectedParentInternalContext',
-  'PredicateReturnedNonOne',
-  'ContractIdAlreadyDeployed',
-  'ContractMismatch',
-  'MessageDataTooLong',
-  'ArithmeticError',
-  'ContractInstructionNotAllowed',
-  'TransferZeroCoins',
-  'InvalidInstruction',
-  'MemoryNotExecutable',
-  'PolicyIsNotSet',
-  'PolicyNotFound',
-  'TooManyReceipts',
-  'BalanceOverflow',
-  'InvalidBlockHeight',
-  'TooManySlots',
-];
-
-export const PANIC_DOC_URL = 'https://docs.rs/fuel-asm/latest/fuel_asm/enum.PanicReason.html';
