+++ packages/transactions/src/configs.ts
@@ -45,3 +45,54 @@ export const FAILED_ASSERT_SIGNAL = '0xffffffffffff0004';
 export const FAILED_ASSERT_NE_SIGNAL = '0xffffffffffff0005';
 
 export const FAILED_UNKNOWN_SIGNAL = '0x0';
+
+export const PANIC_REASONS = [
+  'UnknownPanicReason',
+  'Revert',
+  'OutOfGas',
+  'TransactionValidity',
+  'MemoryOverflow',
+  'ArithmeticOverflow',
+  'ContractNotFound',
+  'MemoryOwnership',
+  'NotEnoughBalance',
+  'ExpectedInternalContext',
+  'AssetIdNotFound',
+  'InputNotFound',
+  'OutputNotFound',
+  'WitnessNotFound',
+  'TransactionMaturity',
+  'InvalidMetadataIdentifier',
+  'MalformedCallStructure',
+  'ReservedRegisterNotWritable',
+  'InvalidFlags',
+  'InvalidImmediateValue',
+  'ExpectedCoinInput',
+  'EcalError',
+  'MemoryWriteOverlap',
+  'ContractNotInInputs',
+  'InternalBalanceOverflow',
+  'ContractMaxSize',
+  'ExpectedUnallocatedStack',
+  'MaxStaticContractsReached',
+  'TransferAmountCannotBeZero',
+  'ExpectedOutputVariable',
+  'ExpectedParentInternalContext',
+  'PredicateReturnedNonOne',
+  'ContractIdAlreadyDeployed',
+  'ContractMismatch',
+  'MessageDataTooLong',
+  'ArithmeticError',
+  'ContractInstructionNotAllowed',
+  'TransferZeroCoins',
+  'InvalidInstruction',
+  'MemoryNotExecutable',
+  'PolicyIsNotSet',
+  'PolicyNotFound',
+  'TooManyReceipts',
+  'BalanceOverflow',
+  'InvalidBlockHeight',
+  'TooManySlots',
+];
+
+export const PANIC_DOC_URL = 'https://docs.rs/fuel-asm/latest/fuel_asm/enum.PanicReason.html';
